Se agrega puerto en el env, la carpeta config y cambios en las rutas de items
