+++ .env.example
@@ -1,0 +1,2 @@
+# Configución del puerto del servidor
+PORT= puerto numérico del servidor
