+++ package.json
@@ -2,9 +2,10 @@
   "name": "biblioteca-back",
   "version": "1.0.0",
   "description": "Backend  del proyecto de Biblioteca Aulica ISFDyT 166 Tandil",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node --watch app.js"
   },
   "repository": {
     "type": "git",
